+++ src/Commands/Run.php
@@ -3,17 +3,14 @@ declare(strict_types = 1);
 
 namespace Formal\Migrations\Commands;
 
-use Innmind\Server\Control\{
-    Server\Processes,
-    Server\Command,
-    Server\Process\TimedOut,
-    Server\Process\Failed,
-    Server\Process\Signaled,
-    Server\Process\Success,
+use Innmind\Server\Control\Server\{
+    Processes,
+    Command,
 };
 use Innmind\Immutable\{
     Map,
-    Either,
+    Attempt,
+    SideEffect,
 };
 
 final class Run
@@ -21,7 +18,7 @@ final class Run
     private Processes $processes;
     /** @var callable(Reference): (callable(Command): Command) */
     private $configure;
-    /** @var Map<Reference, Either<TimedOut|Failed|Signaled, Success>> */
+    /** @var Map<Reference, Attempt<SideEffect>> */
     private Map $alreayRun;
 
     /**
@@ -37,16 +34,20 @@ final class Run
     }
 
     /**
-     * @return Either<TimedOut|Failed|Signaled, Success>
+     * @return Attempt<SideEffect>
      */
-    public function __invoke(Command|Reference $command): Either
+    public function __invoke(Command|Reference $command): Attempt
     {
         if ($command instanceof Command) {
             return $this
                 ->processes
                 ->execute($command)
-                ->unwrap()
-                ->wait();
+                ->flatMap(
+                    static fn($process) => $process
+                        ->wait()
+                        ->map(SideEffect::identity(...))
+                        ->attempt(static fn($e) => new Failure($e)),
+                );
         }
 
         $result = $this
@@ -57,8 +58,12 @@ final class Run
                 fn() => $this
                     ->processes
                     ->execute(($this->configure)($command)($command->command()))
-                    ->unwrap()
-                    ->wait(),
+                    ->flatMap(
+                        static fn($process) => $process
+                            ->wait()
+                            ->map(SideEffect::identity(...))
+                            ->attempt(static fn($e) => new Failure($e)),
+                    ),
             );
         $this->alreayRun = ($this->alreayRun)($command, $result);